Serve logo relative to executable path (#105)
* Serve logo relative to executable path

We used to server files relative to the working directory.
But this can't be considered reliable. This commit changes the way we 
look for files looking for them relative to the executable path.  Since 
he know how the directories under the executable are strutured, this is 
a save way.

This commit also changes the tests accordingly.

Co-Authored-By: Günther Eberl <guenther@eberl.se>

* -mAdd test for os.Executable() error case

* Fix build

* Don't run test verbose

* Improve tests